Create Arduino Code
This is the Arduino Code used to move the machine with no Matlab dependencies. This code will move the machine to preset angles and wait for input before moving to the next angle. The input is done through "Serial Monitor" on the Arduino IDE, the machine will calibrate then wait for input to move to the next angle. Simply press enter to move to the next angle in the array (if you put a character in the command line the motor will jump two angles as the code will read several inputs).