Create Arduino Code with Matlab Dependence
This Arduino code is designed to run with Matlab inputs. The MATLAB file I have attached in the Git Repository is designed to match with this code. The Arduino code speaks to the system and primarily does all the heavy lifting for the machine where as the MATLAB file is designed to input your angles and tell the machine to start the run.